Make implicit narrowing compound assignments explicit
Compound assignments to variables with type byte, char, short, or
float hide narrowing conversions.

Review-Link: https://gwt-review.googlesource.com/#/c/12330/

+++ dev/core/src/com/google/gwt/dev/jjs/impl/Devirtualizer.java
@@ -438,7 +438,7 @@ public class Devirtualizer {
 
     if (program.getTypeJavaLangString().getImplements().contains(enclosingType)) {
       // If it is an interface implemented by String.
-      possibleTargetTypes |= STRING | HAS_JAVA_VIRTUAL_DISPATCH;
+      possibleTargetTypes |= (byte) (STRING | HAS_JAVA_VIRTUAL_DISPATCH);
     }
 
     /////////////////////////////////////////////////////////////////
